Guard

+++ .profile.d/buildpack-boot.sh
@@ -1,3 +1,5 @@
 export PATH="$PATH:/app/mono/bin:/usr/local/bin:/usr/bin:/bin"
 export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:/app/mono/lib"
-ln -s vendor/mono /app
+if ! [ -L "/app/mono" ]; then
+  ln -s vendor/mono /app
+fi
